apps/simulator/requirements.txt
kafka-python==2.0.2
pyyaml==6.0.1
faker==22.0.0
prometheus-client==0.19.0
